Move azul.plugins.metadata.hca.transform and .aggregate into ...indexer
