+++ src/azul/service/manifest_service.py
@@ -102,7 +102,7 @@ from azul.plugins import (
 from azul.plugins.metadata.hca import (
     FileTransformer,
 )
-from azul.plugins.metadata.hca.transform import (
+from azul.plugins.metadata.hca.indexer.transform import (
     value_and_unit,
 )
 from azul.service import (
