+++ test/test_doctests.py
@@ -29,8 +29,8 @@ import azul.openapi
 import azul.openapi.params
 import azul.openapi.responses
 import azul.openapi.schema
+import azul.plugins.metadata.hca.indexer.transform
 import azul.plugins.metadata.hca.service.contributor_matrices
-import azul.plugins.metadata.hca.transform
 import azul.plugins.repository.tdr
 import azul.service.drs_controller
 import azul.service.manifest_service
@@ -75,7 +75,7 @@ def load_tests(_loader, tests, _ignore):
         azul.openapi.schema,
         azul.plugins.metadata.hca.service.contributor_matrices,
         azul.plugins.repository.tdr,
-        azul.plugins.metadata.hca.transform,
+        azul.plugins.metadata.hca.indexer.transform,
         azul.service.drs_controller,
         azul.service.manifest_service,
         azul.service.repository_controller,
